Image tracking
This code will track an image on your screen for the color you specify. In this sample, I chose blue (255,0,0). It will only work for the darkest blues so more washed blue will not be detected, an easy fix by changing the color range. If no object of the color is within the frame, then no box will appear, if an object is presented of the specified color, a bounding box will appear around the object in frame.